Initial commit with a proof of concept for pepperhash
A pepper based hash generator which can e.g. be used for hashing
sensitive data so that the data doesn't have to be stored in a
database. This does create another weakness in that this works
across a network so this isn't perhaps the best idea but still
something that can be used.